renamed some files to be consistent with the rest
