+++ server/index.js
@@ -1,8 +1,8 @@
 const express = require("express");
 const app = express();
 
-require("./startup/db")();
-require("./startup/routes")(app);
+require("./startup/db.startup")();
+require("./startup/routes.startup")(app);
 
 const port = process.env.PORT || 3000;
 const server = app.listen(port, () =>
