+++ server/tests/unit/routes/admin/posts.routes.test.js
@@ -12,7 +12,7 @@ express.Router = jest.fn().mockReturnValue(router);
 const postsCtrl = require("../../../../controllers/posts.controller");
 
 it("should map each route to its corresponding controller method", () => {
-  const postsRoute = require("../../../../routes/admin/posts.route");
+  const postsRoute = require("../../../../routes/admin/posts.routes");
   expect(router.get).toHaveBeenCalledWith("/", postsCtrl.getPosts);
   expect(router.get).toHaveBeenCalledWith("/:id", postsCtrl.getPost);
   expect(router.post).toHaveBeenCalledWith("/", postsCtrl.addPost);
