+++ server/tests/unit/routes/user/posts.routes.test.js
@@ -9,7 +9,7 @@ express.Router = jest.fn().mockReturnValue(router);
 const postsCtrl = require("../../../../controllers/posts.controller");
 
 it("should map each route to its corresponding controller method", () => {
-  const postsRoute = require("../../../../routes/user/posts.route");
+  const postsRoute = require("../../../../routes/user/posts.routes");
   expect(router.get).toHaveBeenCalledWith("/", postsCtrl.getPosts);
   expect(router.get).toHaveBeenCalledWith("/:id", postsCtrl.getPost);
 });
